graph-dependency-parser/min_train.py
import sys
uJ=zip
uG=set
ui=enumerate
un=None
ub=max
uP=len
uq=str
uj=int
ux=range
uz=True
ud=list
uM=float
from collections import defaultdict
import pydantic
import torch
import torch.nn.functional as F
import numpy as np
import tqdm
import yaml
import wandb
from torch import LongTensor
from torch.nn import Linear,CrossEntropyLoss,Module
from torch.optim import Adam
from transformers import AutoTokenizer,RobertaForTokenClassification,RobertaTokenizer,RobertaTokenizerFast,RobertaModel,XLMRobertaTokenizerFast
from ufal.chu_liu_edmonds import chu_liu_edmonds
from datasets import load_dataset
from transformers.modeling_outputs import BaseModelOutputWithPoolingAndCrossAttentions
from util import Config
def strip_none_heads(examples,i):
 u=examples["tokens"][i]
 p=examples["head"][i]
 F=examples["deprel"][i]
 I=[(t,h,d)for t,h,d in uJ(u,p,F)if h!="None"]
 return uJ(*I)
def map_first_occurrence(nums):
 O=uG()
 return{num:i for i,num in ui(nums)if num is not un and num not in O and not O.add(num)}
def pad_to_same_size(lists,padding_symbol):
 S=ub([uP(l)for l in lists])
 return[l+(padding_symbol)*(S-uP(l))for l in lists]
def train(config:Config,UD_DATASET:uq,W:Module):
 def accuracy(uQ,uA,ignore_index=config.ignore_index)->(uj,uj):
  Q=torch.sum(uA!=ignore_index)
  A=torch.sum(uA==uQ)
  return uj(A),uj(Q)
 def tokenize_and_align_labels(examples,skip_index=config.ignore_index):
  r,examples_heads,examples_deprels=[],[],[]
  for D in ux(uP(examples["tokens"])):
   tt,hh,dd=strip_none_heads(examples,D)
   r.append(tt)
   examples_heads.append(hh)
   examples_deprels.append(dd)
  N=B(r,truncation=uz,is_split_into_words=uz,padding=uz)
  U=[]
  R=[]
  T=[]
  ur:ud[uj]=[]
  V=0
  for D,annotated_heads in ui(examples_heads):
   F=examples_deprels[D]
   s=N.word_ids(batch_index=D)
   J=map_first_occurrence(s)
   G=un
   n:ud[uj]=[]
   deprel_ids_here:ud[uj]=[]
   tokens_representing_word_here:ud[uj]=[0]
   for i,word_idx in ui(s):
    if word_idx is un:
     n.append(skip_index)
     deprel_ids_here.append(skip_index)
    elif word_idx!=G:
     if annotated_heads[word_idx]=="None":
      print("A 'None' head survived!")
      sys.exit(0)
     else:
      b=uj(annotated_heads[word_idx])
      P=0 if b==0 else J[b-1]
      n.append(P)
      deprel_ids_here.append(d[F[word_idx]])
      tokens_representing_word_here.append(i)
    else:
     n.append(skip_index)
     deprel_ids_here.append(skip_index)
    G=word_idx
   U.append(n)
   R.append(deprel_ids_here)
   T.append(tokens_representing_word_here)
   ur.append(uP(tokens_representing_word_here))
   if uP(tokens_representing_word_here)>V:
    V=uP(tokens_representing_word_here)
  for q in T:
   q+=[-1]*(V-uP(q))
  N["head"]=U
  N["deprel_ids"]=R
  N["tokens_representing_words"]=T
  N["num_words"]=ur
  N["tokenid_to_wordid"]=[N.word_ids(batch_index=i)for i in ux(uP(examples_heads))]
  return N
 import os
 os.environ["TOKENIZERS_PARALLELISM"]="false"
 if torch.cuda.is_available():
  j="cuda:"+os.getenv("CUDA",default="0")
  x=torch.device(j)
  print("Running on CUDA device "+j)
 elif torch.backends.mps.is_available()and torch.backends.mps.is_built():
  x="mps"
  print("Running on MPS.")
 else:
  x="cpu"
  print("Running on CPU.")
  print("If you're on a Mac, check that you have MacOS 12.3+, an MPS-enabled chip, and current Pytorch.")
 z=["acl","acl:relcl","advcl","advcl:relcl","advmod","advmod:emph","advmod:lmod","amod","appos","aux","aux:pass","case","cc","cc:preconj","ccomp","clf","compound","compound:lvc","compound:prt","compound:redup","compound:svc","conj","cop","csubj","csubj:outer","csubj:pass","dep","det","det:numgov","det:nummod","det:poss","discourse","dislocated","expl","expl:impers","expl:pass","expl:pv","fixed","flat","flat:foreign","flat:name","goeswith","iobj","list","mark","nmod","nmod:poss","nmod:tmod","nsubj","nsubj:outer","nsubj:pass","nummod","nummod:gov","obj","obl","obl:agent","obl:arg","obl:lmod","obl:tmod","orphan","parataxis","punct","reparandum","root","vocative","xcomp","det:predet","obl:npmod","nmod:npmod"]
 d={rel:i for i,rel in ui(z)}
 M=load_dataset("universal_dependencies",UD_DATASET)
 B=XLMRobertaTokenizerFast.from_pretrained("xlm-roberta-base",add_prefix_space=uz)
 m=M.map(tokenize_and_align_labels,batched=uz,batch_size=config.batchsize)
 def print_instances(instances):
  def deprel_or_skip(deprel_id):
   return "--" if deprel_id==-100 else z[deprel_id]
  for i in ux(uP(instances["input_ids"])):
   u=B.convert_ids_to_tokens(instances["input_ids"][i])
   p=instances["head"][i]
   F=[deprel_or_skip(d)for d in instances["deprel_ids"][i]]
   e=instances["tokenid_to_wordid"][i]
   for i,t,h,d in uJ(ux(uP(u)),u,p,F):
    a=e[i]if i<uP(e)else "--"
 W=W.to(x)
 w=['input_ids','attention_mask','head','deprel_ids','tokens_representing_words','num_words','tokenid_to_wordid']
 l=m["train"]
 l.set_format(type='torch',columns=w)
 o=torch.utils.data.DataLoader(l,batch_size=config.batchsize)
 y=m["validation"]
 y.set_format(type='torch',columns=w)
 H=torch.utils.data.DataLoader(y,batch_size=config.batchsize)
 k=CrossEntropyLoss(ignore_index=config.ignore_index,reduction="mean")
 h=(config.betas[0],config.betas[1])
 f=Adam(W.parameters(),lr=config.learning_rate,betas=h)
 X=config.model_dump()|{"betas":h,"dataset":UD_DATASET}
 wandb.init(project=config.wandb_project,config=X)
 wandb.watch(W,log='gradients',log_freq=5)
 for v in ux(config.epochs):
  t=0.0
  W.train()
  for C in tqdm.tqdm(o,desc="Training, Epoch "+uq(v)):
   E=C["input_ids"].to(x)
   g=C["attention_mask"].to(x)
   c=W(E,g)
   p=C["head"].to(x)
   K=k(c,p)
   wandb.log({"batch_loss":K})
   f.zero_grad()
   K.backward()
   f.step()
   if config.debug:
    break
  W.eval()
  with torch.no_grad():
   L=0
   up=0
   uF=0
   uI=0
   uO=0
   uS=0
   for C in tqdm.tqdm(H,desc="Evaluating"):
    E=C["input_ids"].to(x)
    g=C["attention_mask"].to(x)
    c=W(E,g)
    uQ=torch.argmax(c,dim=1).view(-1)
    uA=C["head"].to(x).view(-1)
    A,Q=accuracy(uQ,uA)
    L+=uj(A)
    up+=uj(Q)
    uF+=uj(uQ.shape[0])
    for i in ux(c.shape[0]):
     ur=C["num_words"][i]
     T=C["tokens_representing_words"][i]
     T=torch.narrow(T,0,0,ur.item())
     uD=C["tokenid_to_wordid"][i]
     def lookup(tokid):
      if tokid==0:
       return 0
      else:
       return uD[tokid]+1
     uN=c[i]
     uN=uN[T,:]
     uN=uN[:,T]
     uN=uN.transpose(0,1)
     uN=F.log_softmax(uN,dim=1)
     uU,tree_score=chu_liu_edmonds(uN.numpy(force=uz).astype(np.double))
     uU=uU[1:]
     uR=C['head'][i]
     uR=uR[T].numpy(force=uz)[1:]
     uT=np.vectorize(lookup)
     uR=uT(uR)
     uO+=(uR==uU).sum()
     uS+=uR.shape[0]
     uI+=1
    if config.debug:
     break
   uV=uM(L)/up
   us=uM(uO)/uS
   wandb.log({"dev root accuracy":uV,"dev UAS":us})
 wandb.finish()
# Created by pyminifier (https://github.com/liftoff/pyminifier)

