graph-dependency-parser/requirements.txt
transformers
datasets
conllu
ufal.chu_liu_edmonds
icecream
wandb
pydantic
